+++ src/test/java/org/morphix/reflection/PrimitivesTest.java
@@ -20,12 +20,14 @@ import static org.morphix.reflection.Primitives.fromPrimitive;
 import static org.morphix.reflection.Primitives.isUnboxable;
 import static org.morphix.reflection.Primitives.toPrimitive;
 
+import java.lang.reflect.InvocationTargetException;
 import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.morphix.lang.JavaObjects;
 
 /**
  * Test class for {@link Primitives}.
@@ -58,7 +60,10 @@ class PrimitivesTest {
 
 	@Test
 	void shouldThrowExceptionOnCallingConstructor() {
-		assertThrows(ReflectionException.class, () -> Constructors.IgnoreAccess.newInstance(Primitives.class));
+		ReflectionException reflectionException = assertThrows(ReflectionException.class, () -> Constructors.IgnoreAccess.newInstance(Primitives.class));
+		InvocationTargetException invocationTargetException = JavaObjects.cast(reflectionException.getCause());
+		UnsupportedOperationException unsupportedOperationException = JavaObjects.cast(invocationTargetException.getCause());
+		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
 	}
 
 	@ParameterizedTest